Factor out a few helpers.
