+++ packages/board-server/src/server/boards/invoke.ts
@@ -5,53 +5,13 @@
  */
 
 import { getDataStore } from "@breadboard-ai/data-store";
-import AgentKit from "@google-labs/agent-kit/agent.kit.json" with { type: "json" };
-import {
-  asRuntimeKit,
-  createLoader,
-  inflateData,
-  type KitManifest,
-} from "@google-labs/breadboard";
+import { createLoader, inflateData } from "@google-labs/breadboard";
 import { run, type HarnessRunResult } from "@google-labs/breadboard/harness";
-import { fromManifest } from "@google-labs/breadboard/kits";
-import Core from "@google-labs/core-kit";
-import GeminiKit from "@google-labs/gemini-kit";
-import JSONKit from "@google-labs/json-kit";
-import TemplateKit from "@google-labs/template-kit";
-import { secretsKit } from "../proxy/secrets.js";
-import { getStore } from "../store.js";
 import type { ApiHandler, BoardParseResult } from "../types.js";
 import { BoardServerProvider } from "./utils/board-server-provider.js";
-
-const kits = [
-  secretsKit,
-  asRuntimeKit(Core),
-  asRuntimeKit(JSONKit),
-  asRuntimeKit(TemplateKit),
-  asRuntimeKit(GeminiKit),
-  fromManifest(AgentKit as KitManifest),
-];
-
-const formatRunError = (e: unknown) => {
-  if (typeof e === "string") {
-    return e;
-  }
-  if (e instanceof Error) {
-    return e.message;
-  }
-  if ("message" in (e as any)) {
-    return (e as { message: string }).message;
-  }
-  // Presume it's an ErrorObject.
-  const error = (e as { error: unknown }).error;
-  if (typeof error === "string") {
-    return error;
-  }
-  if (error instanceof Error) {
-    return error.message;
-  }
-  return JSON.stringify(error);
-};
+import { createKits } from "./utils/create-kits.js";
+import { formatRunError } from "./utils/format-run-error.js";
+import { verifyKey } from "./utils/verify-key.js";
 
 export const invoke = async (
   url: string,
@@ -65,7 +25,7 @@ export const invoke = async (
 
   const runner = run({
     url,
-    kits,
+    kits: createKits(),
     loader: createLoader([new BoardServerProvider(path)]),
     store,
     inputs: { model: "gemini-1.5-flash-latest" },
@@ -93,20 +53,6 @@ export const invoke = async (
   return {
     $error: "Run completed without signaling end or error.",
   };
-};
-
-const verifyKey = async (inputs: Record<string, any>) => {
-  const key = inputs.$key;
-  if (!key) {
-    return { success: false, error: "No key supplied" };
-  }
-  const store = getStore();
-  const userStore = await store.getUserStore(key);
-  if (!userStore.success) {
-    return { success: false, error: userStore.error };
-  }
-  delete inputs.$key;
-  return { success: true };
 };
 
 const invokeHandler: ApiHandler = async (parsed, req, res, body) => {
