+++ packages/board-server/src/server/boards/utils/format-run-error.ts
@@ -1,0 +1,26 @@
+/**
+ * @license
+ * Copyright 2024 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export const formatRunError = (e: unknown) => {
+  if (typeof e === "string") {
+    return e;
+  }
+  if (e instanceof Error) {
+    return e.message;
+  }
+  if ("message" in (e as any)) {
+    return (e as { message: string }).message;
+  }
+  // Presume it's an ErrorObject.
+  const error = (e as { error: unknown }).error;
+  if (typeof error === "string") {
+    return error;
+  }
+  if (error instanceof Error) {
+    return error.message;
+  }
+  return JSON.stringify(error);
+};
